created 'user._' files and allowed for database connection
Deleted 'user.js' file so that it can be split into 'user.controller.js' and 'user.route.js'. This separates the class construction and exporting of the api to the index file within api. (which was also changed to include these exports). From there, the 'index.js' file within 'src' folder grabs these exports, and is able to get/push these requests.