added md5 capabilities, and also updated the models for the database
See comments within the code for a more in-depth explanation than the one provided here.

The models are meant to mirror those in the google drive folder. From there, a library called "blueimp-md5" was installed so that we could use md5 for encryption of the passwords (hashing). I am in the midst of adding functionality to the searching within the database. The rest of the changes in the code are simply comments.