Updated where database is grabbing the information from. Additionally added functionality to grab data from the collections titled "puzzles" and "users"
The changes to all these files are to allow for queries to be searched to find data within the db corresponding to the search.

The Url's are working as intended and correctly grab the data and display it.

There are 3 URL's and functions associated with each collection. The first grabs all data within the collection. The second searches within the collection. The third creates data and inserts it into the collection.